backend/requirements.txt
SpeechRecognition
PyPDF2
chromadb
fastapi
google
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-genai
gTTS==2.3.2
langchain-core
langchain-google-genai
langchain-openai
langchain-tavily
langchain_community
langchain-text-splitters
langchainhub
langgraph
langsmith
lxml
pdfplumber
pygame
pyaudio
pypdf
python-dotenv
pyttsx3
tavily-python
uvicorn
google-cloud-texttospeech